+++ Models/BudgetCosts.cs
@@ -7,75 +7,75 @@ namespace Alliance_for_Life.Models
         [Key]
         public int BudgetInvoiceId { get; set; }
 
-        public int ASalandWages { get; set; }
+        public decimal ASalandWages { get; set; }
 
-        public int AEmpBenefits { get; set; }
+        public decimal AEmpBenefits { get; set; }
 
-        public int AEmpTravel { get; set; }
+        public decimal AEmpTravel { get; set; }
 
-        public int AEmpTraining { get; set; }
+        public decimal AEmpTraining { get; set; }
 
-        public int AOfficeRent { get; set; }
+        public decimal AOfficeRent { get; set; }
 
-        public int AOfficeUtilities { get; set; }
+        public decimal AOfficeUtilities { get; set; }
 
-        public int AFacilityIns { get; set; }
+        public decimal AFacilityIns { get; set; }
 
-        public int AOfficeSupplies { get; set; }
+        public decimal AOfficeSupplies { get; set; }
 
-        public int AEquipment { get; set; }
+        public decimal AEquipment { get; set; }
 
-        public int AOfficeCommunications { get; set; }
+        public decimal AOfficeCommunications { get; set; }
 
-        public int AOfficeMaint { get; set; }
+        public decimal AOfficeMaint { get; set; }
 
-        public int AConsulting { get; set; }
+        public decimal AConsulting { get; set; }
 
-        public int SubConPayCost { get; set; }
+        public decimal SubConPayCost { get; set; }
 
-        public int BackgrounCheck { get; set; }
+        public decimal BackgrounCheck { get; set; }
 
-        public int Other { get; set; }
+        public decimal Other { get; set; }
 
-        public int AJanitorServices { get; set; }
+        public decimal AJanitorServices { get; set; }
 
-        public int ADepreciation { get; set; }
+        public decimal ADepreciation { get; set; }
 
-        public int ATechSupport { get; set; }
+        public decimal ATechSupport { get; set; }
 
-        public int ASecurityServices { get; set; }
+        public decimal ASecurityServices { get; set; }
 
-        public double ATotCosts { get; set; }
+        public decimal ATotCosts { get; set; }
 
-        public double AdminFee { get; set; }
+        public decimal AdminFee { get; set; }
         
-        public int Trasportation { get; set; }
+        public decimal Trasportation { get; set; }
 
-        public int JobTraining { get; set; }
+        public decimal JobTraining { get; set; }
 
-        public int TuitionAssistance { get; set; }
+        public decimal TuitionAssistance { get; set; }
 
-        public int ContractedResidential { get; set; }
+        public decimal ContractedResidential { get; set; }
 
-        public int UtilityAssistance { get; set; }
+        public decimal UtilityAssistance { get; set; }
 
-        public int EmergencyShelter { get; set; }
+        public decimal EmergencyShelter { get; set; }
 
-        public int HousingAssistance { get; set; }
+        public decimal HousingAssistance { get; set; }
 
-        public int Childcare { get; set; }
+        public decimal Childcare { get; set; }
 
-        public int Clothing { get; set; }
+        public decimal Clothing { get; set; }
 
-        public int Food { get; set; }
+        public decimal Food { get; set; }
 
-        public int Supplies { get; set; }
+        public decimal Supplies { get; set; }
 
-        public int RFO { get; set; }
+        public decimal RFO { get; set; }
 
-        public double BTotal { get; set; }
+        public decimal BTotal { get; set; }
 
-        public double Maxtot { get; set; }
+        public decimal Maxtot { get; set; }
 
         // Navigation Properties
