+++ Models/ParticipationService.cs
@@ -7,39 +7,39 @@ namespace Alliance_for_Life.Models
         [Key]
         public int PSId { get; set; }
 
-        public double BackgroudCheck { get; set; }
+        public decimal BackgroudCheck { get; set; }
 
-        public double PTranspotation { get; set; }
+        public decimal PTranspotation { get; set; }
 
-        public double PJobTrain { get; set; }
+        public decimal PJobTrain { get; set; }
 
-        public double PEducationAssistance { get; set; }
+        public decimal PEducationAssistance { get; set; }
 
-        public double PBirthCerts { get; set; }
+        public decimal PBirthCerts { get; set; }
         
-        public double PResidentialCare { get; set; }
+        public decimal PResidentialCare { get; set; }
 
-        public double PUtilities { get; set; }
+        public decimal PUtilities { get; set; }
 
-        public double PHousingEmergency { get; set; }
+        public decimal PHousingEmergency { get; set; }
 
-        public double PHousingAssistance { get; set; }
+        public decimal PHousingAssistance { get; set; }
 
-        public double PChildCare { get; set; }
+        public decimal PChildCare { get; set; }
 
-        public double PClothing { get; set; }
+        public decimal PClothing { get; set; }
 
-        public double PFood { get; set; }
+        public decimal PFood { get; set; }
 
-        public double PSupplies { get; set; }
+        public decimal PSupplies { get; set; }
 
-        public double POther { get; set; }
+        public decimal POther { get; set; }
 
-        public double POther2 { get; set; }
+        public decimal POther2 { get; set; }
 
-        public double POther3 { get; set; }
+        public decimal POther3 { get; set; }
 
-        public double PTotals { get; set; }
+        public decimal PTotals { get; set; }
 
         //Navigation Properties
         public Region Region { get; set; }
